+++ lib/src/kotlin_style_extension.dart
@@ -2,14 +2,14 @@ extension FebuKotlinStyleExtension<T> on T? {
   /// exact copy of kotlin like `let` operator. If called on not nil value inner function will be invoked with current value
   /// Be sure to use optional chaining when expecting possible null values
   /// ```dart
-  /// 1.let((int value) => print(value == 1)) // true
-  /// null.let((int value) => print(value == 1)) // ... nothing :)
+  /// 1.let((int value) => value == 1) // true
+  /// null.let((int value) => value == 1) // ... nothing :)
   /// ```
-  T? let(T Function(T) func) {
+  R? let<R>(R Function(T) func) {
     if (this != null) {
       return func(this!);
     }
-    return this;
+    return null;
   }
 
   /// exact copy of kotlin like `apply` operator. If called on not nil value inner function will be invoked
@@ -18,19 +18,19 @@ extension FebuKotlinStyleExtension<T> on T? {
   /// 1.let(() => print('Meeeoow')) // Meeeoow
   /// null.let(() => print('Puuurr')) // ... nothing :)
   /// ```
-  T? apply(T Function() func) {
+  T? apply(void Function(T) func) {
     if (this != null) {
-      return func();
+      func(this as T);
     }
     return this;
   }
 
-  /// exact copy of kotlin like `let` operator. If called on not nil value inner function will be invoked with current value
+  /// exact copy of kotlin like `also` operator. If called on not nil value inner function will be invoked with current value
   /// Be sure to use optional chaining when expecting possible null values
   /// Result of inner function invocation will be returned
   /// ```dart
-  /// 1.also((int value) => 2) // 2
-  /// null.also((int value) => 2) // ... nothing :)
+  /// 1.also((int? value) => value) // 2
+  /// null.also((int? value) => value) // null :)
   /// ```
   R also<R>(R Function(T?) func) {
     return func(this);
@@ -43,7 +43,9 @@ extension FebuKotlinStyleExtension<T> on T? {
   /// 1.run(() => 'Meeeoow') // Meeeoow
   /// null.run(() => 'Puuurr') // ... nothing :)
   /// ```
-  R run<R>(R Function() func) {
-    return func();
+  void run(Function() func) {
+    if (this != null) {
+      func();
+    }
   }
 }